# src/Kentico.Xperience.K13Ecommerce/StoreApi/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Kentico Xperience 13 Store API",
    "version": "v1"
  },
  "paths": {
...
    "/api/store/customer/addresses": {
      "get": {
        "tags": [
          "Customer"
        ],
        "operationId": "CustomerAddresses",
        "parameters": [
          {
            "name": "customerId",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/KAddress"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/store/order/list": {
      "get": {
        "tags": [
          "Order"
        ],
        "operationId": "OrderList",
        "parameters": [
          {
            "name": "Page",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "PageSize",
            "in": "query",
            "schema": {
              "maximum": 100,
              "minimum": 1,
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "OrderBy",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/KOrder"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "KAddress": {
        "required": [
          "addressCity",
          "addressCountryId",
          "addressLine1",
          "addressZip"
        ],
        "type": "object",
        "properties": {
          "addressId": {
            "type": "integer",
            "format": "int32"
          },
          "addressCustomerId": {
            "type": "integer",
            "format": "int32"
          },
          "addressLine1": {
            "minLength": 1,
            "type": "string"
          },
          "addressLine2": {
            "type": "string",
            "nullable": true
          },
          "addressCity": {
            "minLength": 1,
            "type": "string"
          },
          "addressZip": {
            "minLength": 1,
            "type": "string"
          },
          "addressPhone": {
            "type": "string",
            "nullable": true
          },
          "addressName": {
            "type": "string",
            "nullable": true
          },
          "addressPersonalName": {
            "type": "string",
            "nullable": true
          },
          "addressCountryId": {
            "type": "integer",
            "format": "int32"
          },
          "addressStateId": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
...
      "KCustomer": {
        "required": [
          "customerFirstName",
          "customerLastName"
        ],
        "type": "object",
        "properties": {
          "customerId": {
            "type": "integer",
            "format": "int32"
          },
          "customerFirstName": {
            "minLength": 1,
            "type": "string"
          },
          "customerLastName": {
            "minLength": 1,
            "type": "string"
          },
          "customerCompany": {
            "type": "string",
            "nullable": true
          },
          "customerEmail": {
            "type": "string",
            "nullable": true
          },
          "customerPhone": {
            "type": "string",
            "nullable": true
          },
          "customerFax": {
            "type": "string",
            "nullable": true
          },
          "customerUserId": {
            "type": "integer",
            "format": "int32"
          },
          "customerOrganizationId": {
            "type": "string",
            "nullable": true
          },
          "customerTaxRegistrationId": {
            "type": "string",
            "nullable": true
          },
          "customerIsRegistered": {
            "type": "boolean",
            "readOnly": true
          }
        },
        "additionalProperties": false
      },
...
      "KOrder": {
        "type": "object",
        "properties": {
          "orderId": {
            "type": "integer",
            "format": "int32"
          },
          "orderTotalTax": {
            "type": "number",
            "format": "decimal"
          },
          "orderTaxSummary": {
            "type": "string",
            "nullable": true
          },
          "orderInvoiceNumber": {
            "type": "string",
            "nullable": true
          },
          "orderCurrencyCode": {
            "type": "string",
            "nullable": true
          },
          "orderShippingOption": {
            "$ref": "#/components/schemas/KShippingOption"
          },
          "orderPaymentOption": {
            "$ref": "#/components/schemas/KPaymentOption"
          },
          "orderNote": {
            "type": "string",
            "nullable": true
          },
          "orderCustomer": {
            "$ref": "#/components/schemas/KCustomer"
          },
          "orderStatus": {
            "$ref": "#/components/schemas/KOrderStatus"
          },
          "orderGrandTotal": {
            "type": "number",
            "format": "decimal"
          },
          "orderGrandTotalInMainCurrency": {
            "type": "number",
            "format": "decimal"
          },
          "orderTotalPrice": {
            "type": "number",
            "format": "decimal"
          },
          "orderTotalPriceInMainCurrency": {
            "type": "number",
            "format": "decimal"
          },
          "orderTotalShipping": {
            "type": "number",
            "format": "decimal"
          },
          "orderDate": {
            "type": "string",
            "format": "date-time"
          },
          "orderIsPaid": {
            "type": "boolean"
          },
          "orderCulture": {
            "type": "string",
            "nullable": true
          },
          "orderDiscounts": {
            "type": "string",
            "nullable": true
          },
          "orderOtherPayments": {
            "type": "string",
            "nullable": true
          },
          "orderCouponCodes": {
            "type": "string",
            "nullable": true
          },
          "orderPaymentFormattedResult": {
            "type": "string",
            "nullable": true
          },
          "orderTrackingNumber": {
            "type": "string",
            "nullable": true
          },
          "orderSiteId": {
            "type": "integer",
            "format": "int32"
          },
          "orderLastModified": {
            "type": "string",
            "format": "date-time"
          },
          "orderCustomData": {
            "type": "object",
            "additionalProperties": {
              "nullable": true
            },
            "nullable": true
          },
          "orderBillingAddress": {
            "$ref": "#/components/schemas/KAddress"
          },
          "orderShippingAddress": {
            "$ref": "#/components/schemas/KAddress"
          },
          "orderCompanyAddress": {
            "$ref": "#/components/schemas/KAddress"
          },
          "orderItems": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/KOrderItem"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "KOrderItem": {
        "type": "object",
        "properties": {
          "orderItemId": {
            "type": "integer",
            "format": "int32"
          },
          "orderItemUnitPrice": {
            "type": "number",
            "format": "decimal"
          },
          "orderItemTotalPrice": {
            "type": "number",
            "format": "decimal"
          },
          "orderItemTotalPriceInMainCurrency": {
            "type": "number",
            "format": "decimal"
          },
          "orderItemOrderId": {
            "type": "integer",
            "format": "int32"
          },
          "orderItemSkuName": {
            "type": "string",
            "nullable": true
          },
          "orderItemSkuId": {
            "type": "integer",
            "format": "int32"
          },
          "orderItemUnitCount": {
            "type": "integer",
            "format": "int32"
          },
          "orderItemGuid": {
            "type": "string",
            "format": "uuid"
          },
          "orderItemParentGuid": {
            "type": "string",
            "format": "uuid"
          },
          "orderItemBundleGuid": {
            "type": "string",
            "format": "uuid"
          },
          "orderItemSendNotification": {
            "type": "boolean"
          },
          "orderItemText": {
            "type": "string",
            "nullable": true
          },
          "orderItemProductDiscounts": {
            "type": "string",
            "nullable": true
          },
          "orderItemDiscountSummary": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "KOrderStatus": {
        "type": "object",
        "properties": {
          "statusId": {
            "type": "integer",
            "format": "int32"
          },
          "statusName": {
            "type": "string",
            "nullable": true
          },
          "statusDisplayName": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "KPaymentOption": {
        "type": "object",
        "properties": {
          "paymentOptionId": {
            "type": "integer",
            "format": "int32"
          },
          "paymentOptionName": {
            "type": "string",
            "nullable": true
          },
          "paymentOptionDisplayName": {
            "type": "string",
            "nullable": true
          },
          "paymentOptionDescription": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "KShippingOption": {
        "type": "object",
        "properties": {
          "shippingOptionId": {
            "type": "integer",
            "format": "int32"
          },
          "shippingOptionName": {
            "type": "string",
            "nullable": true
          },
          "shippingOptionDisplayName": {
            "type": "string",
            "nullable": true
          },
          "shippingOptionDescription": {
            "type": "string",
            "nullable": true
          },
          "shippingOptionThumbnailUrl": {
            "type": "string",
            "nullable": true
          },
          "shippingOptionPrice": {
            "type": "number",
            "format": "decimal"
          }
        },
        "additionalProperties": false
      },
...
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": { }
      },
...
    },
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "description": "Please enter a valid token",
        "scheme": "Bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "Bearer": [ ]
    }
  ]
}